requirements.txt
# Core dependencies for FLIR One Python library
numpy>=1.20.0
opencv-python>=4.5.0
libusb1>=2.0.0
